feat: :sparkles: Restructure code
